+++ fabtools/nginx.py
@@ -7,8 +7,6 @@ This module provides tools for managing Nginx config files.
 """
 from pipes import quote
 
-from fabric.api import *
-
 from fabtools.files import is_link
 from fabtools.utils import run_as_root
 
@@ -31,8 +29,8 @@ def enable(config):
 
     if not is_link(link_filename):
         run_as_root("ln -s %(config_filename)s %(link_filename)s" % {
-                'config_filename': quote(config_filename),
-                'link_filename': quote(link_filename),
+            'config_filename': quote(config_filename),
+            'link_filename': quote(link_filename),
         })