+++ fabtools/require/files.py
@@ -13,7 +13,7 @@ import os
 from tempfile import mkstemp
 from urlparse import urlparse
 
-from fabric.api import *
+from fabric.api import hide, put, run, settings
 
 from fabtools.files import is_file, is_dir, md5sum
 from fabtools.utils import run_as_root
@@ -103,7 +103,7 @@ def file(path=None, contents=None, source=None, url=None, md5=None,
             path = os.path.basename(urlparse(url).path)
 
         if not is_file(path) or md5 and md5sum(path) != md5:
-            func('wget --progress=dot:mega %(url)s' % locals())
+            func('wget --progress=dot:mega %(url)s -O %(path)s' % locals())
 
     # 3) A local filename, or a content string, is specified
     else:
@@ -135,11 +135,7 @@ def file(path=None, contents=None, source=None, url=None, md5=None,
                 (verify_remote and
                     md5sum(path, use_sudo=use_sudo) != digest.hexdigest())):
             with settings(hide('running')):
-                if source:
-                    put(source, path, use_sudo=use_sudo)
-                else:
-                    put(tmp_file.name, path, use_sudo=use_sudo)
-                    os.remove(tmp_file.name)
+                put(source, path, use_sudo=use_sudo)
 
         if t is not None:
             os.unlink(source)