+++ fabtools/require/rpm.py
@@ -8,8 +8,13 @@ and repositories.
 """
 from __future__ import with_statement
 
-from fabtools.rpm import *
-from fabtools.system import get_arch
+from fabric.api import hide, settings
+from fabtools.rpm import (
+    install,
+    is_installed,
+    uninstall,
+)
+from fabtools.system import get_arch, distrib_release
 from fabtools.utils import run_as_root
 
 
@@ -110,16 +115,22 @@ def repository(name):
     if name == 'rpmforge' and arch == 'i386':
         arch = 'i686'
     supported = {
-      'rpmforge': {'%(arch)s' % locals(): {
-        '6': '%(rpmforge_url)s-%(rpmforge_version)s.el6.rf.i686.rpm' % locals(),
-        '5': '%(rpmforge_url)s-%(rpmforge_version)s.el5.rf.x86_64.rpm' % locals()},
-      'epel': { '%(arch)s' % locals(): {
-        '6': '%(epel_url)s/6/%(arch)s/epel-release-%(epel_version)s.noarch.rpm' % locals(),
-        '5': '%(epel_url)s/5/%(arch)s/epel-release-%(epel_version)s.noarch.rpm' % locals()}}
-        }}
+        'rpmforge': {
+            '%(arch)s' % locals(): {
+                '6': '%(rpmforge_url)s-%(rpmforge_version)s.el6.rf.i686.rpm' % locals(),
+                '5': '%(rpmforge_url)s-%(rpmforge_version)s.el5.rf.x86_64.rpm' % locals(),
+            },
+        },
+        'epel': {
+            '%(arch)s' % locals(): {
+                '6': '%(epel_url)s/6/%(arch)s/epel-release-%(epel_version)s.noarch.rpm' % locals(),
+                '5': '%(epel_url)s/5/%(arch)s/epel-release-%(epel_version)s.noarch.rpm' % locals(),
+            }
+        },
+    }
     keys = {
-    'rpmforge': 'http://apt.sw.be/RPM-GPG-KEY.dag.txt',
-    'epel': '%(epel_url)s/RPM-GPG-KEY-EPEL-%(release)s' % locals()
+        'rpmforge': 'http://apt.sw.be/RPM-GPG-KEY.dag.txt',
+        'epel': '%(epel_url)s/RPM-GPG-KEY-EPEL-%(release)s' % locals(),
     }
     repo = supported[name][str(arch)][str(release)]
     key = keys[name]