+++ fabtools/tests/fabfiles/openvz.py
@@ -2,7 +2,7 @@ from __future__ import with_statement
 
 import time
 
-from fabric.api import *
+from fabric.api import cd, local, put, run, settings, sudo, task
 
 
 @task
@@ -116,7 +116,7 @@ def setup_firewall():
         interfaces=interfaces,
         policy=policy,
         masq=masq,
-        )
+    )
 
     require.shorewall.started()
 
@@ -124,11 +124,13 @@ def setup_firewall():
 def setup_containers():
 
     from fabtools import require
-    from fabtools.openvz import guest
+    from fabtools.openvz import guest, list_ctids
     from fabtools.require.openvz import container
+    from fabtools.system import distrib_family
     import fabtools
 
-    require.deb.package('vzctl')
+    if distrib_family() == 'debian':
+        require.deb.package('vzctl')
 
     NAME = 'debian'
     TEMPLATE = 'debian-6.0-x86_64'
@@ -163,7 +165,7 @@ def setup_containers():
             assert sudo('whoami', user='nobody') == 'nobody'
 
         # Check put
-        with guest(name):
+        with guest(NAME):
             local('echo "toto" > /tmp/toto')
             put('/tmp/toto', '/tmp/toto')
             assert run('test -f /tmp/toto').succeeded
@@ -202,11 +204,11 @@ def setup_containers():
             assert fabtools.files.is_file('/etc/redis/test.conf')
             assert run('echo PING | /opt/redis-2.4.15/redis-cli') == 'PONG'
 
-    assert 'debian' in openvz.list_ctids()
+    assert 'debian' in list_ctids()
 
     # Stop and destroy container
     with container(NAME, TEMPLATE, hostname=NAME, ipadd=IPADD) as ct:
         ct.stop()
         ct.destroy()
 
-    assert 'debian' not in openvz.list_ctids()
+    assert 'debian' not in list_ctids()