+++ fabtools/tests/fabfiles/shorewall.py
@@ -1,4 +1,4 @@
-from fabric.api import *
+from fabric.api import task
 
 
 @task
@@ -17,8 +17,10 @@ def firewall():
             fabtools.shorewall.HTTP(),
             fabtools.shorewall.HTTPS(),
             fabtools.shorewall.SMTP(),
-            fabtools.shorewall.rule(port=1234,
-                source=fabtools.shorewall.hosts(['example.com'])),
+            fabtools.shorewall.rule(
+                port=1234,
+                source=fabtools.shorewall.hosts(['example.com']),
+            ),
         ]
     )