+++ docs/CHANGELOG.rst
@@ -1,10 +1,41 @@
 Changelog
 =========
 
+Version 0.14.0dev (unreleased)
+------------------------------
+
+* Upgrade default pip version to 1.3.1
+* Add support for Vagrant 1.1 providers in functional tests
+* Add ``fabtools.system.cpus`` to get the host's CPU count
+* Less verbose output
+* Move OS detection functions to ``fabtools.system``
+* Better support for Red Hat based systems
+
+Version 0.13.0 (2013-03-15)
+---------------------------
+
+* Add support for managing remote git repositories (Andreas Kaiser)
+* Add intersphinx to docs (Andreas Kaiser)
+* Add HTTP proxy support to speed up functional tests
+* Upgrade default Node.js version to 0.10.0
+* Upgrade default Redis version to 2.6.11
+* Upgrade default Oracle JDK version to 7u17-b02
+* Fix vagrant support (thanks to Dominique Lederer and anentropic)
+
+Version 0.12.0 (2013-03-04)
+---------------------------
+
+* Do not create home directory for system users
+* Fix ``pkg.is_installed`` on SmartOS (thanks to Anthony Scalisi)
+* Fix ``system.get_arch`` (thanks to Kamil Chmielewski)
+* Add support for installing Oracle JDK (thanks to Kamil Chmielewski)
+* Add support for creating Postgres schemas (thanks to Michael Bommarito)
+* Fix ``mysql.user_exists`` (thanks to Serge Travin)
+
 Version 0.11.0 (2013-02-15)
 ---------------------------
 
-* Fix requiring an existing user (thanks to JonPeel)
+* Fix requiring an existing user (thanks to Jonathan Peel)
 * Upgrade default Redis version to 2.6.10
 * Upgrade default Node.js version to 0.8.19
 * Better support for remote hosts where sudo is not installed